requirements.txt
Django==2.2.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==6.2.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
